Render home hero on the server to avoid banner flash
The homepage was a client component, so it first painted the default
translated hero and only swapped in the saved banner after a client-side
fetch resolved — a brief flash of the old banner. Convert the page to a
server component that reads the banner from the database and renders the
hero in the initial HTML, and move the review lists into a small client
component (HomeReviews). The now-unused public banner API route is
removed since the page reads Prisma directly.
